backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Pillow==10.1.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.25.1

# PyJWT==2.8.0
# pytz==2023.3.post1
